+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 3.9-beta
 - Add FileInput() function in Inputs.php
 - Add FileExtensionWhiteList() function in FileUpload.fnc.php
 - Move flag icons from assets/flags/ to their corresponding locale/[code].utf8/ folder
+- Move Help_*.php files from root to their corresponding locale/[code].utf8/ folder
 
 Changes in 3.8
 --------------